Fix test.

+++ matgl/data/tests/test_transformer.py
@@ -12,6 +12,6 @@ class TestNormalizer:
         transformer = Normalizer.from_data(data)
         scaled = transformer.transform(data)
 
-        assert scaled.mean(), pytest.approx(0)
+        assert float(scaled.mean()), pytest.approx(0)
         inverse = transformer.inverse_transform(scaled)
         assert inverse == pytest.approx(data)
